services/order/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-jose==3.3.0
asyncpg==0.28.0
psycopg2-binary==2.9.7
redis==5.0.0
faststream==0.2.5
faststream[kafka]==0.2.5
kafka-python==2.0.2
fast-depends==2.2.0
aiohttp==3.8.5
httpx==0.24.1
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.0.0
stripe==4.2.0
